refactor(price)!: excise the dead sequential price-fetch path
Production price fetching goes through race_current_price, which races
every source concurrently over one transport and takes the first answer.
The older sequential path survived beside it with no production caller:
fetch_current_price defaulted to Gemini, fetch_current_price_from named
a single source, PriceSource::next carried a rotation no caller
advanced, and PriceList::update_current_price wrapped the fetch for the
deprecated lock-holding LightWallet::update_current_price, whose only
callers were two zingolib tests. A wallet.rs doc-comment still directed
callers to the dead path.

This change removes the whole graph: the four zingo-price functions,
the deprecated LightWallet method, its unused PriceError import, and
the rotation test. The live-network smoke test retargets to the race,
and race_current_price now derives each URL from PriceSource::url
instead of repeating the constant pairs.

The two zingolib tests that exercised the deprecated method pinned the
transport-leg error contract. The black-holed-proxy timeout case was
already covered by zingo-price's own tests; the dead-proxy connect
case moves there beside it, so the contract keeps real-socket coverage
at the layer that owns the mechanism.
