+++ zingolib/src/lightclient.rs
@@ -1191,9 +1191,9 @@ mod tests {
     /// clearnet fallback. The route pre-flight variants
     /// (`PriceFetchRequiresMixnet`, `MixnetNotReady::{Unattached, Bootstrapping,
     /// Died}`) pair with `nym::route`'s own `resolve_route` tests; the
-    /// tests here pin the surface wiring and the transport leg. The
-    /// clearnet default lives on [`LightClient::update_current_price`] and
-    /// carries no mixnet contract.
+    /// tests here pin the surface wiring. The transport-leg contract
+    /// (typed connect and timeout failures with their cause chains) is
+    /// pinned in `zingo-price`'s own tests, beside the mechanism.
     #[cfg(feature = "nym")]
     mod price_fetch_contract {
         use crate::lightclient::LightClient;
@@ -1262,101 +1262,6 @@ mod tests {
                 client.mixnet_route(),
                 Ok(crate::nym::MixnetRoute::Clearnet)
             ));
-        }
-
-        /// A dead proxy endpoint surfaces as the typed request variant with
-        /// the [`zingo_net_diag::NetOpFailure`] record beside the reqwest
-        /// source, preserved whole, so a consumer distinguishes a connect
-        /// failure from a TLS or HTTP failure by fields instead of parsing
-        /// prose. The same value converts into
-        /// [`LightClientError::PriceError`] by `From`, which is the exact
-        /// wiring `LightClient::update_current_price`'s `?`
-        /// uses.
-        #[tokio::test]
-        #[allow(deprecated)] // the lock-holding path is the unit under test
-        async fn dead_proxy_failure_is_typed_with_its_source_intact() {
-            use zingo_net_diag::NetOpStage;
-
-            // Bind then drop, so the port is closed when the fetch dials it.
-            let closed = {
-                let listener = std::net::TcpListener::bind("127.0.0.1:0").unwrap();
-                listener.local_addr().unwrap()
-            };
-
-            let mut wallet = wallet();
-            let error = wallet
-                .update_current_price(Some(&closed.to_string()))
-                .await
-                .expect_err("a closed port cannot serve a price");
-
-            match &error {
-                crate::wallet::error::PriceError::PriceError(
-                    zingo_price::PriceError::RequestFailed { failure, source },
-                ) => {
-                    assert!(
-                        source.is_connect(),
-                        "the reqwest failure must keep its kind: {source}"
-                    );
-                    assert!(
-                        matches!(
-                            failure.stage,
-                            NetOpStage::LocalProxyConnect
-                                | NetOpStage::SocksHandshake
-                                | NetOpStage::RemoteConnect
-                        ),
-                        "a dead local proxy must classify as a connect-phase stage: {failure}"
-                    );
-                    assert!(
-                        !failure.cause_chain.is_empty(),
-                        "the cause chain arrives as a vector of layers"
-                    );
-                }
-                other => panic!("the transport failure must arrive typed: {other}"),
-            }
-
-            let surfaced = LightClientError::from(error);
-            assert!(
-                matches!(surfaced, LightClientError::PriceError(_)),
-                "the API surface must report the same typed variant: {surfaced}"
-            );
-        }
-
-        /// A black-holed proxy (the TCP connect completes in the kernel's
-        /// backlog, but no SOCKS5 reply ever comes) resolves within the
-        /// client bound as a typed timed-out failure (net-diag acceptance
-        /// criteria 2 and 9) — the unbounded five-minute field hang can no
-        /// longer happen. Paused tokio time auto-advances the client
-        /// timeout, so the test does not spend the real twenty seconds.
-        #[tokio::test(start_paused = true)]
-        #[allow(deprecated)] // the lock-holding path is the unit under test
-        async fn black_holed_proxy_times_out_typed_within_the_bound() {
-            use zingo_net_diag::NetOpStage;
-
-            let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-            let addr = listener.local_addr().unwrap();
-
-            let mut wallet = wallet();
-            let error = wallet
-                .update_current_price(Some(&addr.to_string()))
-                .await
-                .expect_err("a silent proxy cannot serve a price");
-
-            match &error {
-                crate::wallet::error::PriceError::PriceError(
-                    zingo_price::PriceError::RequestFailed { failure, source },
-                ) => {
-                    assert!(
-                        matches!(failure.stage, NetOpStage::TimedOut { .. }),
-                        "the hang must arrive as a typed timeout stage: {failure}"
-                    );
-                    assert!(
-                        source.is_timeout(),
-                        "the reqwest source must keep its timeout kind: {source}"
-                    );
-                }
-                other => panic!("the timeout must arrive typed: {other}"),
-            }
-            drop(listener);
         }
     }
