+++ zingolib/src/wallet.rs
@@ -22,10 +22,6 @@ use zingo_price::PriceList;
 use crate::config::{ChainType, WalletConfig};
 use crate::data::proposal::ZingoProposal;
 use error::{KeyError, WalletError};
-// The one PriceError-returning method is nym-gated (the mixnet-only price
-// rule), so its import follows the feature.
-#[cfg(feature = "nym")]
-use error::PriceError;
 use keys::unified::{UnifiedAddressId, UnifiedKeyStore};
 
 pub mod error;
@@ -404,38 +400,6 @@ impl LightWallet {
         } else {
             Ok(None)
         }
-    }
-
-    /// Update, record, and return the current price of ZEC.
-    ///
-    /// Currently only USD is supported. When `socks5_proxy` is `Some`, the
-    /// fetch is routed through that local SOCKS5 address (the Nym mixnet
-    /// transport, ADR 0011); `None` fetches over clearnet. The caller resolves
-    /// which route to use.
-    ///
-    /// Deprecated because this method holds `&mut self` — and therefore the
-    /// wallet lock — across the network wait, which is exactly the polling
-    /// blackout the net-diag design removed: every wallet-state observer
-    /// queues behind the fetch. Production surfaces fetch first with
-    /// [`zingo_price::fetch_current_price`] (no lock held) and then record
-    /// the result under a briefly-held lock, as
-    /// `LightClient::update_current_price` does.
-    #[cfg(feature = "nym")]
-    #[deprecated(note = "holds the wallet lock across the network wait; \
-                fetch with zingo_price::fetch_current_price and record with \
-                record_price_update instead")]
-    pub async fn update_current_price(
-        &mut self,
-        socks5_proxy: Option<&str>,
-    ) -> Result<f32, PriceError> {
-        let current_price = self
-            .price_list
-            .update_current_price(socks5_proxy)
-            .await?
-            .price_usd;
-        self.save_required = true;
-
-        Ok(current_price)
     }
 
     /// Records a price fetched *outside* the wallet lock (the net-diag
